Add files via upload
The objective was to write a program that served as a crawling tool for  websites that host political speeches. It is a tad rough on the edges, but it is completely functional and easy to use. Next versions will include a more elegant way to find the initial and ending indexes in the text and eventually, a proper front end and will enable user input.